+++ test/10-http.spec.js
@@ -3,6 +3,7 @@
  */
 
 const { assert } = require('chai');
+const uuid = require('uuid/v4');
 const nock = require('nock');
 const axios = require('./../src/axios');
 const HttpClient = require('./../src/http');
@@ -15,42 +16,34 @@ describe('http', () => {
     const defaultUserAgent = `node-${pkg.name}/${pkg.version}`;
     const customUserAgent = 'custom-ua-123';
 
+    const primaryEndpoint = `http://${uuid()}.example.com`;
+    const defaultUaEndpoint = `http://${uuid()}.example.com`;
+    const customUaEndpoint = `http://${uuid()}.example.com`;
+
 
     /**
      * HTTP mocking
      */
 
     before(() => {
-        nock('http://www.example.com')
-            .persist()
-            .get('/')
-            .reply(200, 'ok');
+        axios.defaults.bypassCustomDnsResolver = true;
 
-        const defaultUaOpts = {
-            reqheaders: {
-                'User-Agent': defaultUserAgent
-            }
-        };
+        const defaultUaOpts = { reqheaders: { 'User-Agent': defaultUserAgent } };
+        const customUaOpts = { reqheaders: { 'User-Agent': customUserAgent } };
 
-        const customUaOpts = {
-            reqheaders: {
-                'User-Agent': customUserAgent
-            }
-        };
+        nock(primaryEndpoint)
+            .persist().get('/').reply(200, 'ok');
 
-        nock('http://default-ua.example.com', defaultUaOpts)
-            .persist()
-            .get('/')
-            .reply(200, 'ok');
+        nock(defaultUaEndpoint, defaultUaOpts)
+            .persist().get('/').reply(200, 'ok');
 
-        nock('http://custom-ua.example.com', customUaOpts)
-            .persist()
-            .get('/')
-            .reply(200, 'ok');
+        nock(customUaEndpoint, customUaOpts)
+            .persist().get('/').reply(200, 'ok');
     });
 
     after(() => {
         axios.defaults.headers.common['User-Agent'] = defaultUserAgent;
+        axios.defaults.bypassCustomDnsResolver = false;
     });
 
 
@@ -68,7 +61,8 @@ describe('http', () => {
      */
 
     it('should http get', async () => {
-        const resp = await testClient.request('http://www.example.com', 'get');
+        const resp = await testClient.request(primaryEndpoint, 'get');
+
         assert.isObject(resp);
         assert.strictEqual(resp.status, 200);
         assert.strictEqual(resp.data, 'ok');
@@ -80,7 +74,7 @@ describe('http', () => {
      */
 
     it('should request using default user-agent', async () => {
-        const resp = await testClient.request('http://default-ua.example.com', 'get');
+        const resp = await testClient.request(defaultUaEndpoint, 'get');
 
         assert.isObject(resp);
         assert.strictEqual(resp.status, 200);
@@ -88,12 +82,12 @@ describe('http', () => {
     });
 
     it('should not request using custom user-agent', async () => {
-        await assert.isRejected(testClient.request('http://custom-ua.example.com', 'get'));
+        await assert.isRejected(testClient.request(customUaEndpoint, 'get'));
     });
 
     it('should request using custom user-agent', async () => {
         axios.defaults.headers.common['User-Agent'] = customUserAgent;
-        const resp = await testClient.request('http://custom-ua.example.com', 'get');
+        const resp = await testClient.request(customUaEndpoint, 'get');
 
         assert.isObject(resp);
         assert.strictEqual(resp.status, 200);
@@ -102,6 +96,6 @@ describe('http', () => {
 
     it('should not request using default user-agent', async () => {
         axios.defaults.headers.common['User-Agent'] = customUserAgent;
-        await assert.isRejected(testClient.request('http://default-ua.example.com', 'get'));
+        await assert.isRejected(testClient.request(defaultUaEndpoint, 'get'));
     });
 });