+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * `fixed` Retry signed API requests on `urn:ietf:params:acme:error:badNonce`
 * `fixed` Minor bugs related to `POST-as-GET` when calling `updateAccount()`
 * `fixed` Ensure subject common name is present in SAN when creating a CSR
+* `fixed` Send empty JSON body when responding to challenges
 
 
 ## v3.1.0 (2019-08-21)