+++ CHANGELOG.md
@@ -1,8 +1,15 @@
+**Notice: [On November 1st, 2019 Let's Encrypt will remove support for unauthenticated GETs from the v2 API](https://community.letsencrypt.org/t/acme-v2-scheduled-deprecation-of-unauthenticated-resource-gets/74380). Please update to `acme-client >= v3.2.0` or `>= v2.3.1` before this date to avoid being affected by this API change.**
+
+
+## v2.3.1 (2019-08-26)
+
+* `backport` Minor bugs related to `POST-as-GET` when calling `updateAccount()`
+* `backport` Send empty JSON body when responding to challenges
+
+
 ## v2.3.0 (2019-08-21)
 
 * `backport` Implemented `POST-as-GET` for all ACME API requests
-
-**Notice: [On November 1st, 2019 Let's Encrypt will remove support for unauthenticated GETs from the v2 API](https://community.letsencrypt.org/t/acme-v2-scheduled-deprecation-of-unauthenticated-resource-gets/74380). Please update to `acme-client >= v3.1.0` or `>= v2.3.0` before this date to avoid being affected by this API change.**
 
 
 ## v2.2.3 (2019-01-25)